fix(RN): broken build after AV pending changes.
