+++ react/features/base/toolbox/components/BaseAudioMuteButton.ts
@@ -5,7 +5,7 @@ import AbstractButton, { IProps } from './AbstractButton';
 /**
  * An abstract implementation of a button for toggling audio mute.
  */
-export default class AbstractAudioMuteButton<P extends IProps, S=any>
+export default class BaseAudioMuteButton<P extends IProps, S=any>
     extends AbstractButton<P, S> {
 
     icon = IconMic;
