+++ react/features/toolbox/components/web/AudioMuteButton.tsx
@@ -2,22 +2,19 @@ import { ClassNameMap, withStyles } from '@mui/styles';
 import React, { ReactElement } from 'react';
 import { connect } from 'react-redux';
 
-import { ACTION_SHORTCUT_TRIGGERED, AUDIO_MUTE, createShortcutEvent } from '../../analytics/AnalyticsEvents';
-import { sendAnalytics } from '../../analytics/functions';
-import { IReduxState } from '../../app/types';
-import { AUDIO_MUTE_BUTTON_ENABLED } from '../../base/flags/constants';
-import { getFeatureFlag } from '../../base/flags/functions';
-import { translate } from '../../base/i18n/functions';
-import { MEDIA_TYPE } from '../../base/media/constants';
-import { IGUMPendingState } from '../../base/media/types';
-import AbstractAudioMuteButton from '../../base/toolbox/components/AbstractAudioMuteButton';
-import AbstractButton, { IProps as AbstractButtonProps } from '../../base/toolbox/components/AbstractButton';
-import { isLocalTrackMuted } from '../../base/tracks/functions';
-import Spinner from '../../base/ui/components/web/Spinner';
-import { registerShortcut, unregisterShortcut } from '../../keyboard-shortcuts/actions';
-import { muteLocal } from '../../video-menu/actions';
-import { SPINNER_COLOR } from '../constants';
-import { isAudioMuteButtonDisabled } from '../functions';
+import { ACTION_SHORTCUT_TRIGGERED, AUDIO_MUTE, createShortcutEvent } from '../../../analytics/AnalyticsEvents';
+import { sendAnalytics } from '../../../analytics/functions';
+import { IReduxState } from '../../../app/types';
+import { translate } from '../../../base/i18n/functions';
+import { IGUMPendingState } from '../../../base/media/types';
+import AbstractButton from '../../../base/toolbox/components/AbstractButton';
+import Spinner from '../../../base/ui/components/web/Spinner';
+import { registerShortcut, unregisterShortcut } from '../../../keyboard-shortcuts/actions';
+import { SPINNER_COLOR } from '../../constants';
+import AbstractAudioMuteButton, {
+    IProps as AbstractAudioMuteButtonProps,
+    mapStateToProps as abstractMapStateToProps
+} from '../AbstractAudioMuteButton';
 
 const styles = () => {
     return {
@@ -32,18 +29,8 @@ const styles = () => {
 /**
  * The type of the React {@code Component} props of {@link AudioMuteButton}.
  */
-interface IProps extends AbstractButtonProps {
+interface IProps extends AbstractAudioMuteButtonProps {
 
-
-    /**
-     * Whether audio is currently muted or not.
-    */
-   _audioMuted: boolean;
-
-   /**
-    * Whether the button is disabled.
-   */
-  _disabled: boolean;
 
   /**
    * The gumPending state from redux.
@@ -62,11 +49,6 @@ interface IProps extends AbstractButtonProps {
  * @augments AbstractAudioMuteButton
  */
 class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
-    accessibilityLabel = 'toolbar.accessibilityLabel.mute';
-    toggledAccessibilityLabel = 'toolbar.accessibilityLabel.unmute';
-    label = 'toolbar.mute';
-    tooltip = 'toolbar.mute';
-    toggledTooltip = 'toolbar.unmute';
 
     /**
      * Initializes a new {@code AudioMuteButton} instance.
@@ -89,10 +71,6 @@ class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
      * @returns {void}
      */
     componentDidMount() {
-        if (typeof APP === 'undefined') {
-            return;
-        }
-
         this.props.dispatch(registerShortcut({
             character: 'M',
             helpDescription: 'keyboardShortcuts.mute',
@@ -107,10 +85,6 @@ class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
      * @returns {void}
      */
     componentWillUnmount() {
-        if (typeof APP === 'undefined') {
-            return;
-        }
-
         this.props.dispatch(unregisterShortcut('M'));
     }
 
@@ -159,13 +133,11 @@ class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
      * @returns {boolean}
      */
     _isAudioMuted() {
-        const { _audioMuted, _gumPending } = this.props;
-
-        if (_gumPending === IGUMPendingState.PENDING_UNMUTE) {
+        if (this.props._gumPending === IGUMPendingState.PENDING_UNMUTE) {
             return false;
         }
 
-        return _audioMuted;
+        return super._isAudioMuted();
     }
 
     /**
@@ -207,26 +179,6 @@ class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
                 </div>
             );
     }
-
-    /**
-     * Changes the muted state.
-     *
-     * @param {boolean} audioMuted - Whether audio should be muted or not.
-     * @protected
-     * @returns {void}
-     */
-    _setAudioMuted(audioMuted: boolean) {
-        this.props.dispatch(muteLocal(audioMuted, MEDIA_TYPE.AUDIO));
-    }
-
-    /**
-     * Return a boolean value indicating if this button is disabled or not.
-     *
-     * @returns {boolean}
-     */
-    _isDisabled() {
-        return this.props._disabled;
-    }
 }
 
 /**
@@ -241,16 +193,11 @@ class AudioMuteButton extends AbstractAudioMuteButton<IProps> {
  * }}
  */
 function _mapStateToProps(state: IReduxState) {
-    const _audioMuted = isLocalTrackMuted(state['features/base/tracks'], MEDIA_TYPE.AUDIO);
-    const _disabled = isAudioMuteButtonDisabled(state);
-    const enabledFlag = getFeatureFlag(state, AUDIO_MUTE_BUTTON_ENABLED, true);
     const { gumPending } = state['features/base/media'].audio;
 
     return {
-        _audioMuted,
-        _disabled,
-        _gumPending: gumPending,
-        visible: enabledFlag
+        ...abstractMapStateToProps(state),
+        _gumPending: gumPending
     };
 }
